ktest/bench: split mod.rs into per-surface files
Mechanical refactor mirroring the unit/ tier's "one file per kernel
surface" rule. bench/mod.rs was 692 lines with 8 bench functions and
their per-bench helper statics co-located; each function is now in its
own file (cap, event, ipc, mm, null, signal, thread, wait_set) and
mod.rs retains only:

- the rule banner and "how to add a benchmark" docstring,
- the shared `cycles_now()` and `log_bench_header()` helpers (now
  `pub(super)`),
- the `run_all()` dispatch.

Each per-surface file owns its own `BENCH_*_STACK` static and entry-fn
where applicable, so a future bench addition is one file edit + one
line in `run_all()`.

No behavioural change.

Refs #24.
